Assignment 2 code
Thank you guys for sending me the code for sum up, I have modify and mix all the part of the code.
But they are having some bugs and problems.
Bug: Every location at the mianpage won't load unless you have press the location once.
Bug: If you drag the date at the viewpage, when you shift back to the main page without drag it back to today, the main page information will be lost.
Problem:Did not use the Tojson func and initialiseFromPDO func (The method of doing the localstorage storing are not good ps.not the structure required in the assignment)

Things remian: The current location. (will be done soon)

Thanks guys. 

Angus